requirements.txt
dghs-imgutils>=0.19.0
torch
huggingface_hub
timm
pillow
pandas
